# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "桃園市婦女生育平均年齡 API",
    "description": "桃園市政府開放資料的年度紀錄；年別使用民國年，平均年齡單位為歲。首次執行會從附帶的 CSV 匯入 SQLite；後續修改保存在 SQLite。",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/birth-ages": {
      "get": {
        "tags": [
          "年度紀錄"
        ],
        "summary": "List Records",
        "operationId": "list_records_api_v1_birth_ages_get",
        "parameters": [
          {
            "name": "year_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 999,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year From"
            }
          },
          {
            "name": "year_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 999,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year To"
            }
          },
          {
            "name": "min_age",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 120,
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Age"
            }
          },
          {
            "name": "max_age",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 120,
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Age"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BirthAgeList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "年度紀錄"
        ],
        "summary": "Create Record",
        "operationId": "create_record_api_v1_birth_ages_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BirthAgeCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BirthAgeRecord"
                }
              }
            }
          },
          "409": {
            "description": "該民國年已有紀錄",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/birth-ages/{roc_year}": {
      "get": {
        "tags": [
          "年度紀錄"
        ],
        "summary": "Get Record",
        "operationId": "get_record_api_v1_birth_ages__roc_year__get",
        "parameters": [
          {
            "name": "roc_year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 999,
              "minimum": 1,
              "title": "Roc Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BirthAgeRecord"
                }
              }
            }
          },
          "404": {
            "description": "該年紀錄不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "年度紀錄"
        ],
        "summary": "Update Record",
        "operationId": "update_record_api_v1_birth_ages__roc_year__put",
        "parameters": [
          {
            "name": "roc_year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 999,
              "minimum": 1,
              "title": "Roc Year"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BirthAgeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BirthAgeRecord"
                }
              }
            }
          },
          "404": {
            "description": "該年紀錄不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "年度紀錄"
        ],
        "summary": "Delete Record",
        "operationId": "delete_record_api_v1_birth_ages__roc_year__delete",
        "parameters": [
          {
            "name": "roc_year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 999,
              "minimum": 1,
              "title": "Roc Year"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "該年紀錄不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BirthAgeCreate": {
        "properties": {
          "roc_year": {
            "type": "integer",
            "maximum": 999.0,
            "minimum": 1.0,
            "title": "Roc Year",
            "description": "民國年，例如 114"
          },
          "average_age": {
            "type": "number",
            "maximum": 120.0,
            "exclusiveMinimum": 0.0,
            "title": "Average Age",
            "description": "當年生育婦女的平均年齡，單位：歲"
          }
        },
        "type": "object",
        "required": [
          "roc_year",
          "average_age"
        ],
        "title": "BirthAgeCreate"
      },
      "BirthAgeList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BirthAgeRecord"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "篩選後的總筆數，不受 limit/offset 影響"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "BirthAgeList"
      },
      "BirthAgeRecord": {
        "properties": {
          "roc_year": {
            "type": "integer",
            "maximum": 999.0,
            "minimum": 1.0,
            "title": "Roc Year",
            "description": "民國年，例如 114"
          },
          "average_age": {
            "type": "number",
            "maximum": 120.0,
            "exclusiveMinimum": 0.0,
            "title": "Average Age",
            "description": "當年生育婦女的平均年齡，單位：歲"
          },
          "gregorian_year": {
            "type": "integer",
            "title": "Gregorian Year",
            "description": "西元年，由民國年加 1911 得出"
          }
        },
        "type": "object",
        "required": [
          "roc_year",
          "average_age",
          "gregorian_year"
        ],
        "title": "BirthAgeRecord"
      },
      "BirthAgeUpdate": {
        "properties": {
          "average_age": {
            "type": "number",
            "maximum": 120.0,
            "exclusiveMinimum": 0.0,
            "title": "Average Age",
            "description": "新的平均年齡，單位：歲；年別為不可修改的資源識別碼"
          }
        },
        "type": "object",
        "required": [
          "average_age"
        ],
        "title": "BirthAgeUpdate"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "tags": [
    {
      "name": "年度紀錄",
      "description": "每一民國年一筆，可新增、查詢、修改及刪除。"
    }
  ]
}